Refactored back PlayerChangedWorldListener.
