+++ src/main/java/me/gnat008/perworldinventory/PerWorldInventory.java
@@ -37,6 +37,7 @@ import me.gnat008.perworldinventory.groups.GroupManager;
 import me.gnat008.perworldinventory.listeners.entity.EntityPortalEventListener;
 import me.gnat008.perworldinventory.listeners.player.PlayerGameModeChangeListener;
 import me.gnat008.perworldinventory.listeners.player.PlayerTeleportListener;
+import me.gnat008.perworldinventory.listeners.player.PlayerChangedWorldListener;
 import me.gnat008.perworldinventory.listeners.player.PlayerQuitListener;
 import me.gnat008.perworldinventory.listeners.player.PlayerSpawnLocationListener;
 import me.gnat008.perworldinventory.listeners.player.PlayerLoginListener;
@@ -172,6 +173,7 @@ public class PerWorldInventory extends JavaPlugin {
         pluginManager.registerEvents(injector.getSingleton(PlayerGameModeChangeListener.class), this);
         pluginManager.registerEvents(injector.getSingleton(PlayerLoginListener.class), this);
         pluginManager.registerEvents(injector.getSingleton(PlayerTeleportListener.class), this);
+        pluginManager.registerEvents(injector.getSingleton(PlayerChangedWorldListener.class), this);
         pluginManager.registerEvents(injector.getSingleton(PlayerQuitListener.class), this);
         pluginManager.registerEvents(injector.getSingleton(EntityPortalEventListener.class), this);
