+++ src/main/java/me/gnat008/perworldinventory/listeners/player/PlayerSpawnLocationListener.java
@@ -23,7 +23,7 @@ import me.gnat008.perworldinventory.config.Settings;
 import me.gnat008.perworldinventory.data.DataWriter;
 import me.gnat008.perworldinventory.groups.Group;
 import me.gnat008.perworldinventory.groups.GroupManager;
-import me.gnat008.perworldinventory.process.TeleportProcess;
+import me.gnat008.perworldinventory.process.InventoryChangeProcess;
 import org.bukkit.Location;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
@@ -37,11 +37,11 @@ public class PlayerSpawnLocationListener implements Listener {
 
     private DataWriter dataWriter;
     private GroupManager groupManager;
-    private TeleportProcess process;
+    private InventoryChangeProcess process;
     private Settings settings;
 
     @Inject
-    PlayerSpawnLocationListener(DataWriter dataWriter, GroupManager groupManager, TeleportProcess process,
+    PlayerSpawnLocationListener(DataWriter dataWriter, GroupManager groupManager, InventoryChangeProcess process,
                                 Settings settings) {
         this.dataWriter = dataWriter;
         this.groupManager = groupManager;
